Refactor: Consolidate results and redirect student results
